backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
pillow>=10.0.0
numpy>=1.24.0
tensorflow>=2.15.0
scikit-learn>=1.3.0
joblib>=1.3.0
pandas>=2.0.0
shap>=0.44.0
opencv-python-headless>=4.9.0
